+++ Generator/RawCopyGenerator.php
@@ -26,7 +26,7 @@ class RawCopyGenerator extends BaseGenerator
     {
         /** @var \SplFileInfo $file */
         foreach (Finder::create()->files()->in($event->getResourcesPath()."raw-copy") as $file) {
-            $relativePath = $relativePath = str_replace($event->getResourcesPath()."raw-copy", "", $file->getRealPath());
+            $relativePath = str_replace($event->getResourcesPath()."raw-copy", "", $file->getRealPath());
             $completePath = $event->getModulePath().$relativePath;
 
             @copy($file->getRealPath(), $completePath);